requirements.txt
# Core ML Libraries
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# MLflow for tracking
mlflow>=2.9.0

# Drift Detection
evidently>=0.4.14

# Model Serving
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0,<3.0.0

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-docker==3.7.0

# Data Processing
pyarrow==14.0.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pytest==7.4.3

